move active code out of bumps.cli so it can be deprecated
